+++ src/llscreen.hh
@@ -53,7 +53,7 @@ public:
 	LLScreen(int16_t resx, int16_t resy, uint8_t depth, bool fullscreen, bool dblbuffer, bool hwsurface);
 	~LLScreen();
 	ifstream *myfopen(string,ios_base::openmode);
-	void print_string(string message, int16_t x, int16_t y, uint8_t ink, uint8_t paper);
+	void print_string(string message, int16_t x, float y, uint8_t ink, uint8_t paper);
 	void set_paletes(bool);
 	void paint_one_pixel(uint8_t value,unsigned char *address);
 	void fullscreen_switch();
